requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.0
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
